third_party/blink/web_tests/virtual/css-custom-state-only-deprecated-syntax/external/wpt/custom-elements/state/state-css-selector-nth-of-expected.txt

This is a testharness.js-based test.
[FAIL] state selector has influence on nth-of when state is applied
  assert_equals: expected "rgb(0, 255, 0)" but got "rgb(255, 0, 0)"
[FAIL] style is invalided on clear()
  assert_equals: expected "rgb(0, 255, 0)" but got "rgb(255, 0, 0)"
Harness: the test ran to completion.
